WIP extract media tiles to component
